requirements.txt
opencv-python
numpy
matplotlib
Pillow
torch
torchvision
seaborn
pandas
